requirements.txt
python-dotenv
asyncio
requests
databases[mysql]
discord.py
discord-components